requirements.txt
asgiref==3.2.10
Django==3.0
djangorestframework==3.10.3
pytz==2020.1
sqlparse==0.3.1
django-cors-headers==3.1.1
django-rest-auth==0.9.5
